LayoutTests/media/media-controller-time-expected.txt

RUN(controller = video.controller)
EVENT(canplaythrough)
RUN(video.play())
RUN(controller.play())
EVENT(playing)
EXPECTED (controller.currentTime > '0') OK
END OF TEST
